FIAP/Segundo_ano/2o_semestre/SAP-GPT/requirements.txt
streamlit
openai
streamlit_extras
opencv-python
scikit-learn
pandas
numpy
tensorflow
keras
torch
ultralytics
pillow


